Abstract effects out of DelegationChain generation
Also:
 - rename Capabilities -> DelegationChain
 - split up into .Types, .Class and DelgationChain module
